Refactored skills list on Resume page
Instead of a boring list with bullet points, I've changed it to a responsive "grid" of images with links and text underneath with the name of the technology.

If possible, I link to my account associated with the technology, such as with GitHub and Bitbucket. Otherwise, I just link to the developer page or most relevant page.